# openapi/aisi.v1.yaml
openapi: 3.0.3
info:
  title: AISI / Dakwah Depok API
  version: 1.0.0
  description: |
    Kontrak SSOT API AISI (Go).
    Envelope sukses: `{ success: true, data, message?, pagination? }`
    Envelope error: `{ success: false, message }`
servers:
  - url: /api/v1
paths:
...
  /evaluations:
    get:
      operationId: listEvaluations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      operationId: createEvaluation
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /evaluations/{id}:
    get:
      operationId: getEvaluation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
    put:
      operationId: updateEvaluation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
  /evaluations/{id}/submit:
    post:
      operationId: submitEvaluation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
  /evaluations/{id}/photos:
    post:
      operationId: uploadEvaluationPhotos
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...